Understand and implement the DTOs and AutoMapper

+++ CwkBooking.Api/Controllers/HotelsController.cs
@@ -1,9 +1,12 @@
-﻿using CwkBooking.Dal;
+﻿using AutoMapper;
+using CwkBooking.Api.DTOs;
+using CwkBooking.Dal;
 using CwkBooking.Domain.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace CwkBooking.Api.Controllers
@@ -17,20 +20,22 @@ namespace CwkBooking.Api.Controllers
         private readonly ILogger<HotelsController> _logger;
         private readonly HttpContext _http;
         private readonly DataContext _ctx;
+        private readonly IMapper _mapper;
         public HotelsController(ILogger<HotelsController> logger, IHttpContextAccessor httpContextAccessor, 
-            DataContext ctx)
+            DataContext ctx, IMapper mapper)
         {
             _logger = logger;
             _http = httpContextAccessor.HttpContext;
             _ctx = ctx;
-
+            _mapper = mapper;
         }
 
         [HttpGet]
         public async Task<IActionResult> GetAllHotels()
         {
             var hotels = await _ctx.Hotels.ToListAsync();
-            return Ok(hotels);
+            var hotelsGet = _mapper.Map<List<HotelGetDTO>>(hotels);
+            return Ok(hotelsGet);
         }
 
 
@@ -39,27 +44,35 @@ namespace CwkBooking.Api.Controllers
         public async Task<IActionResult> GetHotelById(int id)
         {
             var hotel = await _ctx.Hotels.FirstOrDefaultAsync(h => h.HotelId == id);
-            return Ok(hotel);
+
+            if (hotel == null)
+                return NotFound();
+
+            var hotelGet = _mapper.Map<HotelGetDTO>(hotel);
+            return Ok(hotelGet);
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateHotel([FromBody] Hotel hotel)
+        public async Task<IActionResult> CreateHotel([FromBody] HotelCreateDTO hotel)
         {
-            _ctx.Hotels.Add(hotel);
+            var domainHotel = _mapper.Map<Hotel>(hotel);
+
+            _ctx.Hotels.Add(domainHotel);
             await _ctx.SaveChangesAsync();
-            return CreatedAtAction(nameof(GetHotelById), new {id = hotel.HotelId}, hotel);
+
+            var hotelGet = _mapper.Map<HotelGetDTO>(domainHotel);
+
+            return CreatedAtAction(nameof(GetHotelById), new {id = domainHotel.HotelId}, hotelGet);
         }
 
         [HttpPut]
         [Route("{id}")]
-        public async Task<IActionResult> UpdateHotel([FromBody] Hotel updated, int id)
+        public async Task<IActionResult> UpdateHotel([FromBody] HotelCreateDTO updated, int id)
         {
-            var hotel = await _ctx.Hotels.FirstOrDefaultAsync(h => h.HotelId == id);
-            hotel.Stars = updated.Stars;
-            hotel.Description = updated.Description;
-            hotel.Name = updated.Name;
+            var toUpdate = _mapper.Map<Hotel>(updated);
+            toUpdate.HotelId = id;
 
-            _ctx.Hotels.Update(hotel);
+            _ctx.Hotels.Update(toUpdate);
             await _ctx.SaveChangesAsync();
             return NoContent();
         }
@@ -69,6 +82,10 @@ namespace CwkBooking.Api.Controllers
         public async Task<IActionResult> DeleteHotel(int id)
         {
             var hotel = await _ctx.Hotels.FirstOrDefaultAsync(h => h.HotelId == id);
+
+            if (hotel == null)
+                return NotFound();
+
             _ctx.Hotels.Remove(hotel);
             await _ctx.SaveChangesAsync();
             return NoContent();
